Env Variable for percep_ws

+++ docker_scripts/run_devel.sh
@@ -48,20 +48,25 @@ fi
 
 # Enter into the container if it is already running
 if [ "$(docker ps -a --quiet --filter status=running --filter name=$CONTAINER_NAME)" ]; then
-    echo "Attaching to running container: $CONTAINER_NAME"
+    echo -e "\nAttaching to running container: $CONTAINER_NAME"
     docker exec -it $CONTAINER_NAME /bin/bash $@
     exit 0
+fi
+
+# Check if the PERCEP_WS_PATH environment variable is empty
+if [ -z "$PERCEP_WS_PATH" ]; then
+    echo -e "\nThe environment variable : PERCEP_WS_PATH is empty. Point it to the path of the ROS 2 workspace in which the ros-perception-pipeline project is kept !!"
+    exit 1
 fi
 
 # Run the docker container
 docker run --gpus all --shm-size=1g --ulimit memlock=-1 --ulimit stack=67108864 \
 -it --rm --privileged --net=host --ipc=host \
 --name $CONTAINER_NAME \
--v $PWD/../../../src:/root/percep_ws/src \
--v $PWD/../../../models/:/root/percep_ws/models/ \
--v $PWD/ddsconfig.xml:/ddsconfig.xml \
+-v $PERCEP_WS_PATH/src/:/root/percep_ws/src \
+-v $PERCEP_WS_PATH/models/:/root/percep_ws/models/ \
+-v ddsconfig.xml:/ddsconfig.xml \
 --env CYCLONEDDS_URI=/ddsconfig.xml \
 --env="QT_X11_NO_MITSHM=1"  \
 --env="DISPLAY"  \
 object_detection:latest
-
